Cleaning up modules, API types for user account
