+++ src-tauri/src/chat/commands.rs
@@ -4,8 +4,8 @@ use crate::chat::api_types::{APIChatContextNode, APIChatReply, APIChatThread};
 use crate::chat::crud::{create_chat_reply, create_chat_thread, update_reply_sent_to_ai};
 use crate::chat::{ChatContextNode, ChatReplyRow, ChatThreadRow};
 use crate::error::DwataError;
-use crate::store::Store;
 use crate::workspace::helpers::load_ai_integration;
+use crate::workspace::Store;
 use sqlx::query_as;
 use tauri::State;
