+++ src-tauri/src/embedding/commands.rs
@@ -5,8 +5,8 @@ use std::collections::{HashMap, HashSet};
 // use crate::embedding::fastembed::get_embeddings_with_fastembed;
 use crate::embedding::storage::{create_collection_in_qdrant, store_embedding_in_qdrant};
 use crate::error::DwataError;
-use crate::store::Store;
 use crate::workspace::helpers::load_ai_integration;
+use crate::workspace::Store;
 use qdrant_client::qdrant::value::Kind;
 use qdrant_client::qdrant::Value;
 use std::path::PathBuf;
