+++ src-tauri/src/embedding/mod.rs
@@ -1,8 +1,4 @@
 use serde::{Deserialize, Serialize};
-use std::collections::HashSet;
-use std::path::PathBuf;
 
 pub mod commands;
-// pub mod fastembed;
 pub mod storage;
-// mod rust_bert;
