--- src-tauri/src/store/mod.rs
@@ -1,11 +1,0 @@
-use crate::workspace::Config;
-use sqlx::sqlite::SqliteConnection;
-use std::sync::Arc;
-use tokio::sync::Mutex;
-
-pub mod database;
-
-pub struct Store {
-    pub config: Arc<Mutex<Config>>,
-    pub db_connection: Mutex<Option<SqliteConnection>>,
-}
